dashboard/requirements.txt
streamlit==1.44.0
watchfiles==1.0.4
psycopg2-binary==2.9.9